+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/eventdetail/participant/recyclerview/EventParticipantAdapter.kt
@@ -1,10 +1,9 @@
-package com.emmsale.presentation.ui.eventdetail.participant
+package com.emmsale.presentation.ui.eventdetail.participant.recyclerview
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import com.emmsale.databinding.ItemParticipantBinding
 import com.emmsale.presentation.eventdetail.participant.uistate.ParticipantUiState
 
@@ -12,8 +11,8 @@ class EventParticipantAdapter(
     private val requestCompanion: (Long, String) -> Unit,
     private val showMemberProfile: (Long) -> Unit,
 ) : ListAdapter<ParticipantUiState, ParticipantViewHolder>(diffUtil) {
-
     private lateinit var binding: ItemParticipantBinding
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ParticipantViewHolder {
         binding = ItemParticipantBinding.inflate(
             LayoutInflater.from(parent.context),
@@ -38,43 +37,6 @@ class EventParticipantAdapter(
                 oldItem: ParticipantUiState,
                 newItem: ParticipantUiState,
             ): Boolean = oldItem.id == newItem.id
-        }
-    }
-}
-
-class ParticipantViewHolder(
-    private val binding: ItemParticipantBinding,
-    private val requestCompanion: (Long, String) -> Unit,
-    private val showMemberProfile: (Long) -> Unit,
-) : RecyclerView.ViewHolder(binding.root) {
-
-    init {
-        binding.buttonRequestCompanion.setOnClickListener {
-            requestCompanion(
-                binding.participant!!.memberId,
-                binding.participant!!.name,
-            )
-        }
-        binding.ivParticipantImage.setOnClickListener { showMemberProfile(binding.participant!!.memberId) }
-    }
-
-    fun bind(participant: ParticipantUiState) {
-        binding.participant = participant
-    }
-
-    companion object {
-        fun create(
-            parent: ViewGroup,
-            requestCompanion: (Long, String) -> Unit,
-            showMemberProfile: (Long) -> Unit,
-        ): ParticipantViewHolder {
-            val binding = ItemParticipantBinding.inflate(
-                LayoutInflater.from(parent.context),
-                parent,
-                false,
-            )
-
-            return ParticipantViewHolder(binding, requestCompanion, showMemberProfile)
         }
     }
 }
